requirements.txt.txt
streamlit
rembg
pillow
gTTS